feat(filters): quick search filters (#88)
This PR adds quick search filters in the `FilterSelector` menu.

Previously, this menu was used to initially pick a column to create a filter for, then a second menu which is used to mutate the filter value for that column.

From the initial menu, you can now start searching for a "property" (i.e. column option) and if it exists, you can directly set/unset the property.

See here for demo: https://x.com/kianbazza/status/1917701176579219871