fix(filters/ui): fix column option checkbox border visibility in dark mode (#89)
This PR fixes a issue with the `FilterValueOptionController` and `FilterValueMultiOptionController` components which resulted in command item checkboxes (which indicate whether an option is set/unset in the filter) not being visible when unchecked in dark mode.

The issue was that the `--input` border color is the same as the hovered bg color of the command item in dark mode. We now use `--ring` in dark mode for the border.